Fixes instrumenting of cron to preserve environment variables.
This changes the startup script to create a crontab with all of the
supported environment variables prepended to it to ensure they are
available when calling to arkmanager.

This fixes an issue where updating and restarting from cron would fail
to restart the server properly.

This change also includes a number of trailing whitespace trims, as well
as an update to the README where the Variables section had a typo for
the UPDATEONRESTART entry.